Add rcon connection API
This adds a lower level API for rcon connections.

The Minecraft server I am testing against apparently has a rate limit on
requests. Anything over 1 request per second causes the connection to close.
Thus, this API does not allow parallel request on the same TCP connection and
limits them to one per second.

Fragmentation is handled with the assumption that messages are split
exactly on the 4096 byte boundary. This assumption appears to hold with
the server I'm testing against.